feat: Implement accessory management with RecyclerView, form handling, and delete functionality

+++ app/src/main/java/com/example/accessoriesManager/model/Accessory.kt
@@ -9,7 +9,7 @@ data class Accessory(
     @DocumentId
     var id: String = "",
     var name: String = "",
-    var price: Double = 0.0,
+    var price: Long = 0L,
     var createdAt: Timestamp? = null,
     var updatedAt: Timestamp? = null
 )
